Improve extraction metadata from output files.
Searching the entire output file for metadata (bar the QMC table) makes
it hard to have sufficiently unique regular expressions.  This is
particularly the case for certain MPI implementations (hi cx1!) which
add a pile of additional output to STDOUT.  Given we know (because we
can control it) where in the output file a given item of metadata will
appear, we can restrict the search to the relevant section.