Avoid every compilation triggering a cascade.
We wish to set the git sha1 hash at compile time.  As a result,
environment_report.F90 must be recompiled every time any other source
file changes.

It is useful to print out the git sha1 hash and the calculation UUID
(also in environment_report) in other output files.  Unfortunately, this
causes a cascade compilation to be triggered if the compiler updates the
module file every time, even if the interface doesn't change (e.g. if
using ifort rather than gfortran).

A simple fix is to create a copy of this data in calc which is then set
at *run time* from the settings in environment_report.  This drops the
time of the second make in

$ make && touch src/hande_top_level.f90 && time make

from 41s to 5s (almost all of which is spent in the linker) on my
workstation.