Remove D0_population_cycle and proj_energy_cycle.
These are not actually being used in fciqmc/ccmc algorithms any more
following some changes to the update_proj_energy* interfaces a while ago
(possibly by accident?) but it does make the code cleaner and so we just
remove these variables rather than continue to have them cluttering the
place up/restore behaviour that apparently no-one had noticed had
changed.  Further, remove a very confusing comment in ccmc which is
misleading as the proj_energy_cycle estimator was not actually being
used/accumulated!

Also fix a minor bug for the initial status in non-blocking comms, where
the projected estimator was reported to be a factor of ncycles too small.