Add option to recreate the a semi_stoch_t instance
If deterministic states have been moved to a different process then
the semi_stoch_t objects will need altering. The simplest, safest and
most robust way to do this is just to create all the arrays from
scratch, which is usually quick. The exception is the deterministic
space itself - to ensure that it doesn't change we keep the list of
all deterministic states on all processes the same, and use it to
generate list of deterministic states on each process.