Fix propagation of density matrix.
Correct minus sign error due to definition of the death step as being
1-dt(H-S)_{ij} in contrast to 1 + dt (H_T-H)_{ij} when importance
sampling.

Also correct which end we spawn from. This error was missed due to the
usual trial Hamiltonian being the HF Hamiltonian so there was no death
on the diagonal and the minus sign was accounted for in the incorrect
spawning. Picked up when looking at the Free trial density matrix.