Declare Hamiltonian to be symmetric in sparse mode.
When we added the functionality for sparse matrices to not be symmetric
(e.g. as used when distributing the deterministic Hamiltonian in
semi-stochastic), we forgot to explicitly specify that the sparse
Hamiltonian used for Lanczos is stored in a symmetric form, leading to
throwing an (incorrect) run-time error.

Reported by: Marc Coury.