Remove initiator_cas option.
Whilst in the original initiator paper, using a CAS in addition to
a population threshold to define the initiator space has been quietly
dropped from subsequent work.  The population threshold appears to be
the key factor.  As a result, we now just allow the population threshold
to define the initiator space.  This also is more applicable to other
QMC algorithms (ie CCMC and DMQMC).