Use enumerator for DMQMC array indexing
Use an enumerator to create indices to access the various components
of estimator_numerators. estimator_numerators now has a fixed size,
rather than depending on the number of estimates which the user has
asked to sample.

Related to this, the communication of DMQMC estimates has been
cleaned up, by also using an enumerator. This takes the same approach
now used in energy_evaluation. However, DMQMC is a bit more
complicated because there are various arrays of different sizes to
be communicated. So instead, we also create min_ind and max_ind
arrays which hold the minimum and mamximum indices of each quantity
in the array to be communicated.