Use testcode's module extraction functionality.
Where available, use testcode's new ability to import a run-time defined
module and run a python function to obtain the data.  This has several
advantages for us:

* the previous socket-server system to speed up data extraction was
  a massive hack to avoid the pain of import pandas once per file but
  still clunky (and made for some confusing errors if your server was
  using an old version of pyhande even after you updated hande)
* it's about twice as fast as the socket-server system
* the code is much simpler

On the downside, if importlib is not available, we fall back on an
external script to obtain the output.  This no longer does the
socket-server setup so is *much* slower.  But hey, python 2.7 has been
out for almost 5 years!