Sort and compare 32-bit strings in the same way as 64-bit strings.
Rather than comparing the first integer in the bit array and then
subsequent ones until we find different elements, instead compare the
last element first and do a bitwise (rather than arithmetic) comparison.
The order is hence dominated by the most significant non-identical bit
and thus independent of DET_SIZE.  This was not my idea but rather a
great suggestion from Nick Blunt.

Exact (but not statistical) results change as the states are now ordered
differently.

Note this is sufficient for serial QMC calculations to be independent of
DET_SIZE; MPI calculations still have a dependency due to how they
distribute determinants.