Correct benchmark for non-blocking with real amplitudes.
ebb9213b inadvertently fixed an unknown bug in using real amplitudes
with the non-blocking FCIQMC algorithm.

Details:

round_low_population_spawns stochastically rounds particles with weights
less than 1 which are spawned onto previously unoccupied states to
a weight of 1.

Unfortunately it previously only acted upon a (ie the same) global
spawn_t object, even when called from direct_annihilation_received_list,
which is passed explicitly a spawn_t argument.  Undoubtedly we meant to
actually do this on the spawned list received from other processors
instead.  This turns out not to be a disaster as we just insert low
weight determinants into the main list, which are then rounded after
being evolved for an iteration so results are not affected within
statistical accuracy.  Still, it's good to be consistent.