Apply offset to the state hash consistently.
A variable offset is applied to the bit string in CCMC to determine the
processor location in order to allow excitors to gradually move.  We now
apply this in 32-bit chunks so the same data is hashed for DET_SIZE=32
and DET_SIZE=64.

This is sufficient for CCMC calculations to be independent of DET_SIZE.