Add (a copy of) proc_map_t to spawn_t.
Given that spawn_t is the object which holds the particles to be sent to
different processors, it seems sensible that it should know how to send
them.  As such (and in order to later simplify interfaces) we use this
copy instead of qmc_state's version.

WARNING: it is imperative that particles on a given site within the same
particle_t object are sent to the same processor and as such all spawn_t
objects involved in spawning into the same particle_t object *must* have
the same proc_map_t information.  The programmer *must* keep this
up-to-date (e.g. after load balancing).