Don't require scalapack for serial compilation.
Until 73a540f, we included the netlib source for pdlaprnt and pslaprnt
to work around an (old) bug in these procedures in MKL.  This was
removed because the bug has been fixed in MKL for several years, so
no-one should be affected.  It turns out that we never protected calling
these procedures  with an #ifdef, so required them (even though they
couldn't be used) in serial as well.  Adding this guard prevents an
unwanted and unnecessary dependency on scalapack for serial compilation.