Avoid crash when missetting shift in HANDE.COMM
The intuitive way of setting the shift in HANDE.COMM would be to
set it to a single value, but because the shift is an array, it needs
to be set to an array.  This does some checking on the value read by
lua, and allows a subset of the shifts to be set.