Allow all processors to fail running the lua script.
Partially undo 81aec5f: a failure to run the lua script entirely (ie lua
fails to parse the supplied buffer) is sufficiently catastrophic that
all processors should trigger a stop_all.  81aec5f only allowed the
parent to throw the error, which leads to a deadlock if another
processor fails to run the lua script and the parent does not have the
same failure.