Remove the end1 and end2 components from the rdm_t type
The end1 and end2 components of rdm_t were simply used to work
out the bitstrings of an RDM element. There is no reason for
these variables to be stored a part of this (or any) derived
type. They can just be local variables.

This commit therefore removes end1 and end2 from rdm_t and simply
makes them local variables in the two places needed.