Update DMQMC input checks
Update the DMQMC input checks, which includes checks for using reduced
density matrices with non-Heisenberg systems, and checking each estimator
is only used for the Heisenberg model separately (in preparation for other
estimators which work with other systems). Also, check that we are using
a bipartite system with staggered magnetisaton for *all* systems, not
just the Heisenberg model.

Also some wording has been changed slighly for accuracy.

Also, add full stops to comments!