Move the excit_dist_start option to be associated with the find_weights option
Before it was associated with excit_dist. However, the excit_dist option
calculates the excitation distribution at all report loops and doesn't
use excit_dist_start at all, so this made no sense.

Now, excit_dist_start is called find_weights_start in the dmqmc
block, instead of the operators block.

Also, for clarity, rename the 'ground_state_start' option to
'ground_rdm_start', to make it clear that this is to with RDMs,
and clarify the difference between the different options for
averaging in the ground state.