Use the string length returned by flu_tolstring rather than size.
This avoids a seg fault in gfortran 5.2 on allocate(character(size(str) : ...).
It's not obvious to me why it's causing a seg fault (accessing size outside of
the allocate statement is fine according to valgrind but inside results in an
uninitialised read) but it's simple to avoid given we are told the string length
anyway by flu_tolstring.