Use new allowed_excitation flag.
This replaces the old behaviour of checking that a matrix element was
non-zero as a proxy for whether the excitation generator generated a
valid excitation.  This is a slightly brittle check when doing linked
CCMC, as one term in a commutator being small is not indicative of the
magnitude of other terms, so significant terms may be neglected.