Use pkgutil rather than relying on ImportError.
python 3.3 introduced namespace packages which means a directory without
__init__.py can now be imported.  The directory structure:

pyhande/
        pyhande/
                __init__.py

can hence now be imported even if the top-level pyhande/ directory isn't
on sys.path.  We used to use

try:
    import pyhande
except ImportError:
    sys.path.append (...)
    import pyhande

as a way of importing from the local project if pyhande wasn't
installed, which breaks on python 3.3 onwards as the top-level pyhande
is imported as a namespace package.  The fix to restore desired
behaviour is to instead use:

if not pkgutil.find_loader('pyhande'):
    sys.path.append(...)
import pyhande

This works on python 2.6, 2.7 and 3.5 (and probably 3.3 and 3.4).