let autostart ctd if no opt_block in 1st sweep
Sometimes, blocking might always fail in the first number_of_reblockings (i.e. the
first sweep after which it is usually checked whether the minimum is in the
first pos_min_frac of the first sweep data). If blocking has always
failed, a minimum cannot be found and the variable min_index denoting
the index of the current minimum stays -1 as assigned in the
beginning. The check used to be "if min_index < int(pos_min_frac * j)",
continue, minimum found. To avoid entering this if clause if min_index
is -1 still, the additional condition "min_index > -1" was added.

An unhelpful error message if min_index is -1 was deleted.