Fix semi-stochastic initialisation with shift_start_iteration
Enable semi-stochastic projection immediately if shift_start_iteration
is used and the calculation is restarted with the shift already varying.
The iteration offset is only calculated when the shift starts varying
and this is not information we have available when restarting a calculation so
instead just enable semi-stochastic projection immediately.