added check for uncommented output end
This fixes a problem when extracting Monte Carlo data if a comment
without a hash immediately follows the data table in the output file.

Until now we have assumed that the data table is followed by a blank
line which then marks the end of the data table. Hashed comments are
ignored when importing the data table.
An issue arises if the data table is followed by a comment that is
not marked with the hash symbol and does not leave a blank line.
This can for example happen on certain clusters that print summary
lines at the end of an output file which do not start with a hash
symbol. This can be a problem if the run is not exited softly with
HANDE.COMM.
This fix adds another test for the end of the data table which is
checking for non numerical values in lines that do not start with a
hash symbol.