Allocate buffer for deterministic space distribution only on root
When reading the deterministic space in for semi-stochastic, the fix in
3f56aab3affc4ee990cde65c5eeab782bd5d4376 allocated a buffer on all
processors for assigning each determinant to a given processor. However,
it did so using a variable only initialised on the root processor,
leading to crashes on some platforms as we attempted to allocate an
impossible amount of memory. Fortunately we only need (and use!) the
buffer on the root processor, so the fix is to only allocate it on the
root (thus also saving memory).