[WIP] Bug fixes for the the harmonic parameters and the documentation file
This code is still a work in progress [WIP], the push includes these bug fixes to the original code: 'shift_harmonic_critical_damp' was too many characters and so was not recognized; the 'shift_harmonic_forcing' parameter that was calculated from 'shift_harmonic_critical_damp' was not being printed correctly in the JSON block. This push also includes the qmc.rst documentation file with the new parameters added. Now that these bugs are fixed, the next step are test_suite tests. This is not a final version of the code.