Revamp merging with main list.
The maths of the wall-Chebyshev projector suggests that there is no more iterative updating of cluster amplitudes, but at the end of each Chebyshev iteration the cluster amplitudes that were sampled in the foregoing step gets replaced. This is slightly weird and its effect on population dynamics will definitely be tested.