Fix bug in generating the MP1 WF, also clean up code for PR.
The legacy code did something pretty confusing (and also gave more or less correct energies +- 10 mHa, which raised some alarms but the code never broke so), so I never bothered touching it until today when it broke in a system with bit_string_len > 1. Now the MP2 correlation energies are exact.